Extract Commands from Compiler
